examples/sequential-thinking-python/requirements.txt
openai
fastapi
uvicorn[standard]
termcolor
python-dotenv
mcp>=1.10.0
fastmcp>=2.10.0
pytest
pytest-asyncio 